#6: re-organise + add integration tests for flush db
