+++ src/main/java/store/commands/CommandParser.java
@@ -1,0 +1,5 @@
+package store.commands;
+
+public class CommandParser {
+
+}
