+++ src/main/java/store/engine/DiskRecord.java
@@ -1,4 +1,4 @@
-package store;
+package store.engine;
 
 import java.io.IOException;
 import java.nio.ByteBuffer;
